requirements_conda2.txt
# packages in environment at C:\Users\er112\anaconda3\envs\capstone_kobert:
#
# Name                    Version                   Build  Channel
blas                      1.0                         mkl  
bzip2                     1.0.8                h2bbff1b_6  
ca-certificates           2024.3.11            haa95532_0  
certifi                   2024.2.2         py38haa95532_0  
cffi                      1.16.0           py38h2bbff1b_1  
chardet                   3.0.4                    pypi_0    pypi
charset-normalizer        2.0.4              pyhd3eb1b0_0  
colorama                  0.4.6            py38haa95532_0  
cpuonly                   2.0                           0    pytorch-test
cython                    3.0.10                   pypi_0    pypi
expat                     2.6.2                hd77b12b_0  
filelock                  3.13.1           py38haa95532_0  
freetype                  2.12.1               ha860e81_0  
fsspec                    2024.3.1         py38haa95532_0  
gluonnlp                  0.10.0                   pypi_0    pypi
gmpy2                     2.1.2            py38h7f96b67_0  
huggingface_hub           0.20.3           py38haa95532_0  
idna                      2.6                      pypi_0    pypi
intel-openmp              2021.4.0          haa95532_3556  
jinja2                    3.1.3            py38haa95532_0  
jpeg                      9e                   h2bbff1b_1  
lcms2                     2.12                 h83e58a3_0  
lerc                      3.0                  hd77b12b_0  
libdeflate                1.17                 h2bbff1b_1  
libffi                    3.4.4                hd77b12b_1  
libjpeg-turbo             2.0.0                h196d8e1_0  
libpng                    1.6.39               h8cc25b3_0  
libtiff                   4.5.1                hd77b12b_0  
libuv                     1.44.2               h2bbff1b_0  
libwebp-base              1.3.2                h2bbff1b_0  
lz4-c                     1.9.4                h2bbff1b_1  
markupsafe                2.1.3            py38h2bbff1b_0  
mkl                       2021.4.0           haa95532_640  
mkl-service               2.4.0            py38h2bbff1b_0  
mkl_fft                   1.3.1            py38h277e83a_0  
mkl_random                1.2.2            py38hf11a4ad_0  
mpc                       1.1.0                h7edee0f_1  
mpfr                      4.0.2                h62dcd97_1  
mpir                      3.0.0                hec2e145_1  
mpmath                    1.3.0            py38haa95532_0  
mxnet                     1.7.0.post2              pypi_0    pypi
networkx                  3.1              py38haa95532_0  
ninja                     1.10.2               haa95532_5  
ninja-base                1.10.2               h6d14046_5  
numpy                     1.17.4                   pypi_0    pypi
openjpeg                  2.4.0                h4fc8c34_0  
openssl                   3.0.13               h2bbff1b_1  
packaging                 23.2             py38haa95532_0  
pillow                    10.3.0           py38h2bbff1b_0  
pip                       24.0             py38haa95532_0  
pycparser                 2.21               pyhd3eb1b0_0  
python                    3.8.19               h1aa4202_0  
python-graphviz           0.8.4                    pypi_0    pypi
pytorch                   2.3.0               py3.8_cpu_0    pytorch-test
pytorch-mutex             1.0                         cpu    pytorch-test
pyyaml                    6.0.1            py38h2bbff1b_0  
regex                     2023.10.3        py38h2bbff1b_0  
requests                  2.18.4                   pypi_0    pypi
sentencepiece             0.1.95           py38h59b6b97_0  
setuptools                69.5.1           py38haa95532_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.45.3               h2bbff1b_0  
sympy                     1.12             py38haa95532_0  
tbb                       2021.8.0             h59b6b97_0  
tk                        8.6.14               h0416ee5_0  
tokenizers                0.13.3           py38h49fca51_0  
torchaudio                2.3.0                  py38_cpu    pytorch-test
torchvision               0.18.0                 py38_cpu    pytorch-test
tqdm                      4.66.2           py38h9909e9c_0  
transformers              4.24.0           py38haa95532_0  
typing-extensions         4.11.0           py38haa95532_0  
typing_extensions         4.11.0           py38haa95532_0  
tzdata                    2024a                h04d1e81_0  
urllib3                   1.22                     pypi_0    pypi
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.43.0           py38haa95532_0  
xz                        5.4.6                h8cc25b3_1  
yaml                      0.2.5                he774522_0  
zlib                      1.2.13               h8cc25b3_1  
zstd                      1.5.5                hd43e919_2  
